Add ERC20Burnable preset (#252)
* added erc20 burnable preset

* updated burnable with new ownable module

* removed unused import and changed constructor to initializer

* added event testing

* adapt to new dir structure

* remove unused mock

* removed ownable and split up burn from test

* removed lingering owner and bumped version comment
