+++ src/openzeppelin/token/erc20/presets/ERC20Burnable.cairo
@@ -1,11 +1,12 @@
 # SPDX-License-Identifier: MIT
+# OpenZeppelin Cairo Contracts v0.3.1 (token/erc20/presets/ERC20Burnable.cairo)
 
 %lang starknet
 
 from starkware.cairo.common.cairo_builtins import HashBuiltin
+from starkware.cairo.common.uint256 import Uint256
 from starkware.starknet.common.syscalls import get_caller_address
 from starkware.cairo.common.bool import TRUE
-from starkware.cairo.common.uint256 import Uint256
 
 from openzeppelin.token.erc20.library import ERC20
 
@@ -32,8 +33,8 @@ end
 
 @view
 func name{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }() -> (name: felt):
     let (name) = ERC20.name()
@@ -42,8 +43,8 @@ end
 
 @view
 func symbol{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }() -> (symbol: felt):
     let (symbol) = ERC20.symbol()
@@ -52,8 +53,8 @@ end
 
 @view
 func totalSupply{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }() -> (totalSupply: Uint256):
     let (totalSupply: Uint256) = ERC20.total_supply()
@@ -62,8 +63,8 @@ end
 
 @view
 func decimals{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }() -> (decimals: felt):
     let (decimals) = ERC20.decimals()
@@ -72,8 +73,8 @@ end
 
 @view
 func balanceOf{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(account: felt) -> (balance: Uint256):
     let (balance: Uint256) = ERC20.balance_of(account)
@@ -82,8 +83,8 @@ end
 
 @view
 func allowance{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(owner: felt, spender: felt) -> (remaining: Uint256):
     let (remaining: Uint256) = ERC20.allowance(owner, spender)
@@ -91,13 +92,13 @@ func allowance{
 end
 
 #
-# Externals
+# External
 #
 
 @external
 func transfer{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(recipient: felt, amount: Uint256) -> (success: felt):
     ERC20.transfer(recipient, amount)
@@ -106,8 +107,8 @@ end
 
 @external
 func transferFrom{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(
         sender: felt,
@@ -120,8 +121,8 @@ end
 
 @external
 func approve{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(spender: felt, amount: Uint256) -> (success: felt):
     ERC20.approve(spender, amount)
@@ -130,8 +131,8 @@ end
 
 @external
 func increaseAllowance{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(spender: felt, added_value: Uint256) -> (success: felt):
     ERC20.increase_allowance(spender, added_value)
@@ -140,8 +141,8 @@ end
 
 @external
 func decreaseAllowance{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(spender: felt, subtracted_value: Uint256) -> (success: felt):
     ERC20.decrease_allowance(spender, subtracted_value)
@@ -150,11 +151,23 @@ end
 
 @external
 func burn{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
         range_check_ptr
     }(amount: Uint256):
-    let (owner) = get_caller_address()
-    ERC20._burn(owner, amount)
+    let (caller) = get_caller_address()
+    ERC20._burn(caller, amount)
+    return ()
+end
+
+@external
+func burnFrom{
+        syscall_ptr: felt*,
+        pedersen_ptr: HashBuiltin*,
+        range_check_ptr
+    }(account: felt, amount: Uint256):
+    let (caller) = get_caller_address()
+    ERC20._spend_allowance(account, caller, amount)
+    ERC20._burn(account, amount)
     return ()
 end
